backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Libraries
torch==2.1.1
transformers==4.36.2
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.2
prophet==1.1.4
sentence-transformers==2.2.2
langchain==0.0.350
openai==1.3.7

# Data Processing
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
plotly==5.17.0

# Database & Storage
pymongo==4.6.0
redis==5.0.1
pinecone-client==2.2.4

# AWS Services
boto3==1.34.0
botocore==1.34.0

# WebSocket & Real-time
websockets==12.0
python-socketio==5.10.0

# Security & Privacy
cryptography==41.0.8
pycryptodome==3.19.0
diffprivlib==0.6.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
